Fix the serial port startup hangs
The arduino now sends RESET as soon as it boots.

If the script receives this, it calls stty to fix the serial port
parameters, and restarts the BEGIN sequence.

The next time around, the arduino won't reboot on connection, so the
write-rom.py script starts quickly. This will persist until the serial
port is disconnected/reconnected, or something resets the serial port
parameters, like uploading a new binary to the arduino.

Once that happens, we get the RESET again, and fix it again.

This has been bugging me for sooooo long.