If the page wait fails, try verifying it
I'm having trouble with the page wait failing. It reads the wrong data.

I've found that if I verify the page in this case, it reads back
correct.

This change adds some retry logic to the page write. If the page wait
fails, verify the page. If the verify fails, retry the write. After five
retries we bail out.